backend/requirements.txt
langchain
langchain-community
langchain-core
langchain-google-genai
langchain-experimental
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
pinecone
langchain-pinecone
pydantic
pydantic-settings
alembic
pandas
requests
python-multipart
PyPDF2
tabulate
openpyxl       
python-docx

pymysql
cryptography